Use syscall.Dup3 on ARM64 as Dup2 is unsupported
